% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis.server-package.R
\docType{package}
\name{rtemis.server-package}
\alias{rtemis.server}
\alias{rtemis.server-package}
\title{rtemis.server: rtemis Server}
\description{
rtemis Websocket Server to allow rtemislive connections
}
\details{
\pkg{rtemis.server}: Amino Acid Annotation format
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.rtemis.org}
  \item Report bugs at \url{https://github.com/rtemis-org/rtemis.server/issues}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID}) [copyright holder]
}

}
